Dates managed as floats, not singles
git-svn-id: https://sources.ecopath.org/svn/Ecopath/trunk/Ecopath6@3951 a725721c-0ed6-864e-a6a8-02c2a1988c9d

+++ Sources/EwECore/Core/cCore.vb
@@ -2218,7 +2218,7 @@ Public Class cCore
     Friend m_EwEModelUnitMonetaryCustom As String = ""
     Friend m_EwEModelAuthor As String = ""
     Friend m_EwEModelContact As String = ""
-    Friend m_EwEModelLastSaved As Single = 0
+    Friend m_EwEModelLastSaved As Double = 0
     Friend m_EwEModelAreaName As String = ""
     Friend m_EwEModelSouth As Single = 0
     Friend m_EwEModelNorth As Single = 0
